separate chat and home routes
