+++ package.json
@@ -52,6 +52,7 @@
     "react-dom": "^19.0.0",
     "react-hook-form": "^7.54.2",
     "react-resizable-panels": "^2.1.7",
+    "react-router-dom": "^7.4.0",
     "recharts": "^2.15.1",
     "shadcn": "2.4.0-canary.17",
     "sonner": "^2.0.1",
