+++ src/components/Header.tsx
@@ -1,15 +1,18 @@
 import { Avatar, AvatarFallback } from "@/components/ui/avatar";
 import { ModeToggle } from "@/components/ui/mode-toggle";
+import { Button } from "@/components/ui/button";
+import { LogOut } from "lucide-react";
 
 interface HeaderProps {
   userName: string;
+  onLogout?: () => void;
 }
 
-const Header = ({ userName }: HeaderProps) => {
+const Header = ({ userName, onLogout }: HeaderProps) => {
   return (
-    <div className=" w-full border-b bg-background/95 backdrop-blur">
-      <div className=" flex h-14 items-center justify-between px-5">
-        <span className="text-xl">Conversations</span>
+    <div className="w-full border-b bg-background/95 backdrop-blur">
+      <div className="flex h-14 items-center justify-between px-4">
+        <span className="text-lg">Chat Room</span>
 
         <div className="flex items-center gap-2">
           <ModeToggle />
@@ -18,6 +21,17 @@ const Header = ({ userName }: HeaderProps) => {
               <AvatarFallback>{userName.charAt(0).toUpperCase()}</AvatarFallback>
             </Avatar>
             <span className="text-sm font-medium">{userName}</span>
+            {onLogout && (
+              <Button
+                variant="ghost"
+                size="icon"
+                onClick={onLogout}
+                className="ml-1 cursor-pointer"
+                title="Logout"
+              >
+                <LogOut className="h-4 w-4" />
+              </Button>
+            )}
           </div>
         </div>
       </div>
