+++ src/components/UserNameForm.tsx
@@ -19,7 +19,7 @@ const UserNameForm = ({ onSubmit, initialUserName = '' }: UserNameFormProps) => 
   };
 
   return (
-    <div className="h-screen flex items-center justify-center bg-gradient-to-b from-slate-50 to-slate-100 dark:from-slate-900 dark:to-slate-800">
+    <div className="h-screen flex items-center justify-center bg-background">
       <Card className="w-full max-w-md shadow-lg">
         <CardHeader className="space-y-1">
           <CardTitle className="text-2xl font-bold text-center">Welcome to Chat App</CardTitle>
